Update info properly

+++ CNU/LocationService.h
@@ -28,8 +28,8 @@ static bool hasLocation;
 @property (nonatomic, retain) CLLocationManager *locationManager;
 @property (nonatomic, retain) SettingsService *settingsService;
 @property (nonatomic) bool dieFlag;
-@property (nonatomic) long lastUpdate;
-@property (nonatomic) long lastPublishedUpdate;
+@property (nonatomic) long long lastUpdate;
+@property (nonatomic) long long lastPublishedUpdate;
 @property (nonatomic) dispatch_source_t timerSource;
 
 -(id)initWithSettings:(SettingsService *) settings;
